Fix AST accessors always returning first child
AST generation that I copied from jrsonnet turned out to be too naive:
it didn't support AST nodes more than one child of the same type.
Unfortunately, those cases are inherently more difficult:
just returning nth child of a type results in strange cases,
For example, when analyzing binary expression, we want the right hand
sub-expression to not only follow the left one, but also the operator.
For this reason I decided not to try to generate such accessors,
and instead to require target side to supplement them.