Replace let-decl with def-decl
Now `let` is used in let-expressions, and all top-level declarations
have their unique keywords.
Therefore, whenever parser sees `def` it starts a new declaration,
even if there was no semicolon before that.
Consequently, if `let` is encountered during parsing, it is probably
because the user forgot the `in` or to open parens.