Don't assign syntax to internal curried functions
For example in expr `\x y -> x` which translates to (somewhat)
`Lambda(x, Lambda(y, x))` only the external lambda should have syntax
node mapped to itself. Reason is that there is no sensible syntax node
to use anyway: in case if such HIR node is mentioned in an error, the
whole lambda should be displayed.