Implement let-polymorphism
This message ought to be much longer, but summarizing:
We change unification algorithm from rustc's `ena` union-find crate,
where unification happens outside types, in a special table,
into implementation with `Link` type case, which creates somewhat
implicit union-find table.
On top of this we implement OCaml's algorithm described at
https://okmij.org/ftp/ML/generalization.html, which handles let-scoped
generalization of unification variables.

An unfortunate side-effect of this change is that it disables recursion
of definitions. They are order-dependent now. This means that Camlot has
currently no recursion whatsoever, which means it's probably
not Turing-complete. Obviously recursion has to be added later in some
form, even if not as powerful as it has been until now.

This is a big change and may be reverted later.