Improve handling missing colon in type annotation
Treat `(x y` as two params, not param with annotation.
It's a heuristic: such sequence may be a result of deleting a colon
from inside `(x: y` (note that whatever is on the right side, maybe
a right paren, is invisible to the parser), but probably more often
it is a first stage of type annotating `x` out of plain `x y`.